Bug resuelto con contador_período

+++ TPO-Headers/cinematica.h
@@ -1,5 +1,5 @@
 
 #include <RegsLPC1769.h>
 
-bool mover_motor(uint8_t motor , uint8_t pos_ini , uint8_t pos_final , uint8_t velocidad , uint8_t aceleracion );
+bool mover_motor(uint8_t motor , uint32_t pos_ini , uint32_t pos_final , uint32_t velocidad , uint32_t aceleracion );
 //uint8_t mover_motores (uint8_t pos_iniciales[], uint8_t pos_finales[] , uint8_t velocidades[], uint8_t aceleraciones );
